requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
apturl==0.5.2
arrow==1.3.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
aws-lambda-builders==1.45.0
aws-sam-cli==1.107.0
aws-sam-translator==1.83.0
awscli==1.18.69
backports.zoneinfo==0.2.1
bcrypt==3.1.7
binaryornot==0.4.4
blinker==1.7.0
boto3==1.34.20
boto3-stubs==1.34.12
botocore==1.34.20
botocore-stubs==1.34.20
Brlapi==0.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
cfn-lint==0.83.8
chardet==3.0.4
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
colorama==0.4.3
command-not-found==0.3
cookiecutter==2.5.0
cryptography==41.0.7
cupshelpers==1.0
dateparser==1.2.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.11
djangorestframework==3.15.1
docker==7.0.0
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
fasteners==0.14.1
filelock==3.13.3
flask==3.0.0
flatbuffers==24.3.25
fsspec==2024.3.1
future==0.18.2
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
httplib2==0.14.0
huggingface-hub==0.22.2
idna==3.6
importlib-metadata==7.1.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jschema-to-python==1.2.3
jsonpatch==1.33
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
junit-xml==1.9
keras==2.13.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==18.1.1
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.5
mpmath==1.3.0
mypy-boto3-apigateway==1.34.0
mypy-boto3-cloudformation==1.34.0
mypy-boto3-ecr==1.34.0
mypy-boto3-iam==1.34.8
mypy-boto3-kinesis==1.34.0
mypy-boto3-lambda==1.34.0
mypy-boto3-s3==1.34.14
mypy-boto3-schemas==1.34.0
mypy-boto3-secretsmanager==1.34.17
mypy-boto3-signer==1.34.0
mypy-boto3-sqs==1.34.0
mypy-boto3-stepfunctions==1.34.0
mypy-boto3-sts==1.34.0
mypy-boto3-xray==1.34.0
mysql-connector-python==8.3.0
netifaces==0.10.4
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
olefile==0.46
opt-einsum==3.3.0
packaging==24.0
pandas==2.0.3
paramiko==2.6.0
pbr==6.0.0
pexpect==4.6.0
Pillow==7.0.0
pkgutil-resolve-name==1.3.10
protobuf==4.25.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydantic==2.5.3
pydantic-core==2.14.6
pygments==2.17.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==23.3.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-slugify==8.0.1
pytz==2024.1
pyxdg==0.26
PyYAML==6.0.1
referencing==0.32.1
regex==2023.12.25
reportlab==3.5.34
requests==2.31.0
requests-oauthlib==2.0.0
requests-unixsocket==0.2.0
rich==13.7.0
roman==2.0.0
rpds-py==0.17.1
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
safetensors==0.4.2
sarif-om==1.0.4
scikit-learn==1.3.2
scipy==1.10.1
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
systemd-python==234
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.4.0
tokenizers==0.15.2
tomlkit==0.12.3
tqdm==4.66.2
transformers==4.39.3
types-awscrt==0.20.0
types-python-dateutil==2.8.19.20240106
types-s3transfer==0.10.0
typing-extensions==4.5.0
tzdata==2024.1
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==2.2.1
usb-creator==0.3.7
wadllib==1.3.3
watchdog==3.0.0
werkzeug==3.0.2
wrapt==1.16.0
xkit==0.0.0
zipp==3.18.1
